+++ src/modules/deploy/request-deploy.ts
@@ -10,7 +10,7 @@ import { currentVersion, resolveDockerfilePath } from "@/plugins";
 
 import type { StartBuildParams } from "../build";
 import { generateBuildTag } from "../build/generate-build-tag";
-import { stageCommitAndPushAll } from "../git/git-utils";
+import { isUnstagedFiles } from "../git/git-utils";
 import { askForDeployEnvironmentInfo } from "./ask-deploy-environment-info";
 import { parseOptionsToAppConfig } from "./parse-options-to-app-config";
 
@@ -35,6 +35,10 @@ export async function requestDeploy(options: InputOptions) {
 	let dockerFile = resolveDockerfilePath({ targetDirectory, env });
 	if (options.isDebugging) console.log("requestDeploy() > dockerFile :>> ", dockerFile);
 	if (!dockerFile) return;
+
+	// Warn about uncommited files
+	const shouldShowGitWarning = await isUnstagedFiles(options.targetDirectory);
+	if (shouldShowGitWarning) logWarn(`Please stage files & commit before deploying.`);
 
 	/**
 	 * [1] Parse cli options, validate the input params
@@ -79,16 +83,16 @@ export async function requestDeploy(options: InputOptions) {
 	/**
 	 * [4] Stage, commit & push configuration files (dx.json) to GIT repository:
 	 */
-	try {
-		await stageCommitAndPushAll({
-			directory: options.targetDirectory,
-			message: `build(${env}): ${options.buildImage}`,
-		});
-	} catch (e) {
-		// Stop the process if this throws any errors
-		logError(`Can't commit files for building this app: ${e}`);
-		return;
-	}
+	// try {
+	// 	await stageCommitAndPushAll({
+	// 		directory: options.targetDirectory,
+	// 		message: `build(${env}): ${options.buildImage}`,
+	// 	});
+	// } catch (e) {
+	// 	// Stop the process if this throws any errors
+	// 	logError(`Can't commit files for building this app: ${e}`);
+	// 	return;
+	// }
 
 	/**
 	 * [5] Notify the commander & call API to request server build: